requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
langchain_community
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
langchain-groq
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Utilities
bs4
faiss-cpu
pypdf
grandalf